Fix: revert to float16 for RTX A6000 Ampere (sm_86, not Ada)
Ampere tensor cores perform best with FP16, not BF16.
BF16 advantage only applies to Ada Lovelace (sm_89) and Hopper.
SDPA attention kept - still provides fused kernel speedup on sm_86.

+++ embedder.py
@@ -59,9 +59,9 @@ class MuRILEmbedder:
             self.tokenizer = AutoTokenizer.from_pretrained(
                 self.config.embedding_model_name
             )
-            # RTX6000 Ada: load MuRIL in bfloat16 for faster embedding on Ada tensor cores
+            # RTX A6000 Ampere: float16 for MuRIL embedding (Ampere FP16 tensor cores)
             embed_dtype = (
-                torch.bfloat16 if self.device.type == "cuda"
+                torch.float16 if self.device.type == "cuda"
                 else torch.float32
             )
             self.model = AutoModel.from_pretrained(
